Add files via upload
The objective of this project is to analyse the Walmart Sales dataset, with the goal of gaining insights into various aspects of the business, including the performance of different store branches and products, trends in product sales, and customer behaviour. The primary aim is to examine opportunities for enhancing and optimizing sales strategies. The dataset used in this analysis was sourced from the Kaggle Walmart Sales Forecasting Competition.